refactor: replace Loader and LoaderV2 with new LoadingIndicator component
