--- src/components/LoadingIndicator/LoadingIndicator.tsx
@@ -1,7 +1,0 @@
-import styles from "./LoadingIndicator.module.css";
-
-const LoadingIndicator = (): JSX.Element => {
-  return <span className={styles.loadingAnimation} />;
-};
-
-export default LoadingIndicator;
