--- src/components/common/LoaderV2/LoaderV2.module.css
@@ -1,12 +1,0 @@
-.loader {
-  animation: spin 1.5s linear infinite;
-}
-
-@keyframes spin {
-  0% {
-    transform: rotate(0deg);
-  }
-  100% {
-    transform: rotate(360deg);
-  }
-}
