--- src/components/common/LoaderV2/LoaderV2.tsx
@@ -1,7 +1,0 @@
-import styles from "./LoaderV2.module.css";
-
-const LoaderV2 = () => {
-  return <img src="/images/loader.png" className={styles.loader} />;
-};
-
-export default LoaderV2;
