+++ src/components/common/LoadingIndicator/LoadingIndicator.module.css
@@ -70,8 +70,6 @@
 }
 
 .loadingAnimation {
-  font-size: 24px;
-  font-weight: bold;
   position: relative;
 }
