+++ src/components/pages/liquidity-page/components/Pools/Pools.tsx
@@ -3,7 +3,6 @@ import {useRouter} from "next/navigation";
 
 import MobilePools from "@/src/components/pages/liquidity-page/components/Pools/MobilePools/MobilePools";
 import DesktopPools from "@/src/components/pages/liquidity-page/components/Pools/DesktopPools/DesktopPools";
-import LoaderV2 from "@/src/components/common/LoaderV2/LoaderV2";
 import ActionButton from "@/src/components/common/ActionButton/ActionButton";
 import Pagination from "@/src/components/common/Pagination/Pagination";
 import {SearchBar} from "@/src/components/common/SearchBar/SearchBar";
@@ -12,6 +11,7 @@ import useDebounce from "@/src/hooks/useDebounce";
 
 import clsx from "clsx";
 import styles from "./Pools.module.css";
+import LoadingIndicator from "@/src/components/common/LoadingIndicator/LoadingIndicator";
 
 const Pools = () => {
   const router = useRouter();
@@ -86,6 +86,14 @@ const Pools = () => {
         />
       </div>
 
+      {/* Loading State */}
+      {isLoading && (
+        <div className={styles.loadingFallback}>
+          <LoadingIndicator fontSize="mc-mono-xxxxl" />
+          <p>Loading pools...</p>
+        </div>
+      )}
+
       {/* Pools List (Mobile and Desktop) */}
       <MobilePools poolsData={data} orderBy={orderBy} handleSort={handleSort} />
       <DesktopPools
@@ -93,14 +101,6 @@ const Pools = () => {
         orderBy={orderBy}
         handleSort={handleSort}
       />
-
-      {/* Loading State */}
-      {isLoading && (
-        <div className={styles.loadingFallback}>
-          <LoaderV2 />
-          <p>Loading pools...</p>
-        </div>
-      )}
 
       {/* Pagination */}
       {data && data.length > 0 && (
