+++ src/components/pages/points-page/PointsRankTable/PointsRankTable.module.css
@@ -2,6 +2,7 @@
   display: flex;
   flex-direction: column;
   align-items: center;
+  padding: 20px 0;
 }
 
 .table {
