+++ src/components/pages/view-position-page/components/PositionView/ReserveItem/ReserveItem.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import CoinWithAmount from "@/src/components/common/CoinWithAmount/CoinWithAmount";
 import styles from "./ReserveItem.module.css";
 import {useFormattedReserveValue} from "./useFormattedReserveValue";
-import Loader from "@/src/components/common/Loader/Loader";
+import LoadingIndicator from "@/src/components/common/LoadingIndicator/LoadingIndicator";
 
 interface ReserveItemsProps {
   assetId: string;
@@ -34,7 +34,7 @@ const ReserveItem = ({
           )}
         </div>
       ) : (
-        <Loader color="gray" />
+        <LoadingIndicator fontSize="mc-mono-m" />
       )}
     </div>
   );
